just-another-agent/pyproject.toml
[project]
name = "just-another-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Priyanshu Yashwant Deshmukh",email = "69320370+priyansh4320@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "langgraph (>=0.4.1,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=0.3.2,<0.4.0)",
    "uvicorn (>=0.34.2,<0.35.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
